+++ backend/common/util.ts
@@ -48,6 +48,13 @@ export function findFirstAvailableId(ranges: Range[], ids: number[]): number {
     return 0;
 }
 
+/**
+ * Calculates the SHA256 hash of specified content and returns it in specified encoding.
+ * 
+ * @param content Content to hash
+ * @param encoding Encoding to use for output
+ * @returns SHA256 hash of the content encoded as specified
+ */
 export function getSha256(content: string, encoding: "hex" | "base64") {
     const sha256 = crypto.createHash("sha256");
     sha256.update(content);