+++ src/commands/confirm-sync-object-ids.ts
@@ -2,21 +2,23 @@ import { commands, env, Uri, window } from "vscode";
 import { URLS } from "../lib/constants";
 
 const OPTION = {
-    YES: "Yes, overwrite any recorded object ID assignment information in the back end",
-    NO: "No, I don't think I can do it at this time",
-    LEARN: "I am not sure, tell me more about synchronization",
+    UPDATE: "Update. I want to merge actual object ID assignments with what's already recorded.",
+    REPLACE: "Replace. I want to completely replace recorded object ID assignments with actual state.",
+    NO: "Nothing. I have changed my mind, I won't do it at this time.",
+    LEARN: "I am not sure. Tell me more about synchronization.",
 };
 
 export const confirmSyncObjectIds = async () => {
     let result = await window.showQuickPick(Object.values(OPTION), {
-        placeHolder: "Are you sure you want to synchronize your object ID assignment information with the back end?"
+        placeHolder: "How would you like to synchronize object ID assignment information with the back end?"
     });
     switch (result) {
-        case OPTION.YES:
-            commands.executeCommand("vjeko-al-objid.sync-object-ids");
+        case OPTION.REPLACE:
+        case OPTION.UPDATE:
+            commands.executeCommand("vjeko-al-objid.sync-object-ids", { merge: result === OPTION.UPDATE });
             break;
         case OPTION.LEARN:
             env.openExternal(Uri.parse(URLS.SYNCHRONIZATION_LEARN));
-            break;   
+            break;
     }
 };