+++ src/commands/sync-object-ids.ts
@@ -1,14 +1,22 @@
-import { RelativePattern, Uri, window, workspace, WorkspaceFolder } from "vscode";
+import { RelativePattern, Uri, workspace } from "vscode";
 import * as fs from "fs";
 import { ALObject, parseObjects } from "../lib/parser";
 import { getManifest } from "../lib/AppManifest";
-import { Backend, ConsumptionInfo } from "../lib/Backend";
+import { Backend } from "../lib/Backend";
 import { UI } from "../lib/UI";
 import { ALWorkspace } from "../lib/ALWorkspace";
 import { ObjIdConfig } from "../lib/ObjIdConfig";
 import { MeasureTime } from "../lib/MeasureTime";
 import { Output, output } from "../features/Output";
 import { Config } from "../lib/Config";
+import { ConsumptionInfo } from "../lib/BackendTypes";
+import { LABELS } from "../lib/constants";
+
+interface SyncOptions {
+    merge: boolean,
+    skipQuestion: boolean,
+    uri: Uri,
+}
 
 async function getWorkspaceFolderFiles(uri: Uri): Promise<Uri[]> {
     let folderPath: string = uri.fsPath;
@@ -31,7 +39,7 @@ function getObjectDefinitions(uris: Uri[]): ALObject[] {
     return objects;
 }
 
-function getConsumption(objects: ALObject[]): ConsumptionInfo {
+function getActualConsumption(objects: ALObject[]): ConsumptionInfo {
     let consumption: ConsumptionInfo = {};
     for (let object of objects) {
         let { type, id } = object;
@@ -45,11 +53,12 @@ function getConsumption(objects: ALObject[]): ConsumptionInfo {
 /**
  * Synchronizes object ID consumption information with the Azure back end.
  * 
+ * @param patch Flag that indicates whether patching (merge) rather than full replace should be done
  * @param uri Uri of a document or a workspace folder for which to run the synchronization
  * @returns 
  */
-export const syncObjectIds = async (uri?: Uri) => {
-    uri = await ALWorkspace.selectWorkspaceFolder(uri);
+export const syncObjectIds = async (options?: SyncOptions) => {
+    let uri = await ALWorkspace.selectWorkspaceFolder(options?.uri);
     if (!uri) return;
 
     const manifest = getManifest(uri);
@@ -59,10 +68,14 @@ export const syncObjectIds = async (uri?: Uri) => {
         return;
     }
 
-    // Easter egg or cheating, call it what you want... but this is to prevent "creative" users from intentionally breaking the demo for others.
-    if (manifest.id === "c454e488-56ca-4414-bd68-1d3a2548abf2") {
-        UI.sandbox.showSandboxInfo("synchronized");
-        return;
+    let authKey = ObjIdConfig.instance(uri).authKey;
+
+    if (!options?.merge) {
+        let consumption = await Backend.getConsumption(manifest!.id, authKey);
+        if (consumption?._total) {
+            let answer = await UI.sync.showAreYouSure();
+            if (answer === LABELS.SYNC_ARE_YOU_SURE.NO) return;
+        }
     }
 
     output.log("Starting syncing object ID consumption with the back end");
@@ -79,9 +92,9 @@ export const syncObjectIds = async (uri?: Uri) => {
 
     MeasureTime.log("loading", "parsing");
 
-    const consumption: ConsumptionInfo = getConsumption(objects);
+    const consumption: ConsumptionInfo = getActualConsumption(objects);
 
-    if (await Backend.syncIds(manifest?.id, consumption, ObjIdConfig.instance(uri).authKey || "")) {
+    if (await Backend.syncIds(manifest?.id, consumption, !!(options?.merge), ObjIdConfig.instance(uri).authKey || "")) {
         UI.sync.showSuccessInfo();
     }
 }